Fixed logout ; Fixed login response ; Regex ; Error handling
Fixed logout : remove cookie when user is null

Fixed login and signup response : we expected data to be directly accessible but user's data is actually in the "user" object of the response.

More maintainable regex expression : they are now exportable for more consistency.

Error handling : [sign up] handling codes when falling in the catch statement.